test(backend): organize e2e tests
