+++ packages/backend/tests/features/editUser/editUser.e2e.test.ts
@@ -5,8 +5,8 @@ import { UpdateUserResponse, UserInput, UserExceptions } from '@dddforum/shared/
 import { APIClient } from '@dddforum/shared/src/core';
 import { UserInputBuilder } from '@dddforum/shared/tests/support/builders';
 import { GenericErrors } from '@dddforum/shared/src/shared';
-import { DatabaseFixtures } from '../support/fixtures/DatabaseFixtures';
-import { CompositionRoot } from '../../src/core';
+import { DatabaseFixtures } from '../../support/fixtures/DatabaseFixtures';
+import { CompositionRoot } from '../../../src/core';
 
 const feature = loadFeature(path.join(sharedTestRoot, 'features/editUser.feature'));
